# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "alwism.com - API Documentation",
        "description": "This is the API Gateway for [www.alwism.com](https://www.alwism.com).  You can find out more about alwism at [https://www.alwism.com/sitemechanism](https://www.alwism.com/sitemechanism) or visit to github repository [github](http://url.cc/).",
        "contact": {
            "name": "Contact the Developer",
            "url": "https://www.alwism.com/contactme"
        },
        "version": "1.0.3"
    },
    "servers": [
        {
            "url": "api/v1"
        }
    ],
    "paths": {
...
        "/api/admin/siteui/details": {
            "post": {
                "tags": [
                    "siteUi"
                ],
                "summary": "Create or update the existing details",
                "operationId": "570f8768aa42cb33dc95e8ed9b8cc0ab",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Details"
                },
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/language": {
            "post": {
                "tags": [
                    "languages"
                ],
                "summary": "Create a new language",
                "operationId": "db92e01ce98abb18be2deb0304292ba5",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Languages"
                },
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/language/{id}": {
            "get": {
                "tags": [
                    "languages"
                ],
                "summary": "Get language by ID",
                "operationId": "b8d7d853b82234c680bdae553e2f3825",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Detail": {
                "title": "Detail",
                "description": "Detail model",
                "required": [
                    "topic",
                    "description",
                    "name",
                    "email",
                    "phone",
                    "whatsapp",
                    "skype",
                    "address",
                    "experience",
                    "language"
                ],
                "properties": {
                    "topic": {
                        "title": "topic",
                        "description": "About details topic",
                        "type": "string",
                        "format": "string",
                        "default": "Well experienced Software Engineer"
                    },
                    "description": {
                        "title": "description",
                        "description": "About details description",
                        "type": "string",
                        "format": "string",
                        "default": "description"
                    },
                    "name": {
                        "title": "name",
                        "description": "About details name",
                        "type": "string",
                        "format": "string",
                        "default": "Alwis Madhusanka"
                    },
                    "email": {
                        "title": "email",
                        "description": "About details email",
                        "type": "string",
                        "format": "string",
                        "default": "email@example.com"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "About details phone",
                        "type": "string",
                        "format": "string",
                        "default": "+44 0123 45678"
                    },
                    "whatsapp": {
                        "title": "whatsapp",
                        "description": "About details whatsapp",
                        "type": "string",
                        "format": "string",
                        "default": "+44 0123 45678"
                    },
                    "skype": {
                        "title": "skype",
                        "description": "About details skype",
                        "type": "string",
                        "format": "string",
                        "default": "live::skypeid"
                    },
                    "address": {
                        "title": "address",
                        "description": "About details address",
                        "type": "string",
                        "format": "string",
                        "default": "9 Somewhere, Earth"
                    },
                    "experience": {
                        "title": "experience",
                        "description": "About details experience",
                        "type": "integer",
                        "format": "int",
                        "default": "5"
                    },
                    "language": {
                        "title": "language",
                        "description": "About details language",
                        "type": "string",
                        "format": "string",
                        "default": "English, Sinhala"
                    }
                },
                "type": "object"
            },
            "Language": {
                "title": "Language",
                "description": "Language model",
                "required": [
                    "name",
                    "image",
                    "weight"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Language name",
                        "type": "string",
                        "default": "Golang"
                    },
                    "image": {
                        "title": "image",
                        "description": "Language Image",
                        "type": "string",
                        "format": "binary"
                    },
                    "weight": {
                        "title": "weight",
                        "description": "Language weight",
                        "type": "integer",
                        "default": "1"
                    }
                },
                "type": "object"
            },
...
        },
...
        "requestBodies": {
...
            "Details": {
                "description": "Details Request Body",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/Detail"
                        }
                    },
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Detail"
                        }
                    }
                }
            },
            "Languages": {
                "description": "Languages Request Body",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "$ref": "#/components/schemas/Language"
                        }
                    },
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/Language"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}